backend/requirements.txt
# ============================================
# PERSONAL HEALTH RECORD - DEPENDENCIES
# ============================================

# -----------------------------------------
# 🚀 Core Web Framework
# -----------------------------------------
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# -----------------------------------------
# 🗄️ Database
# -----------------------------------------
motor==3.3.2                        # Async MongoDB driver
pymongo==4.6.0                      # MongoDB driver
beanie==1.23.6                      # ODM for MongoDB

# -----------------------------------------
# 🔐 Authentication & Security
# -----------------------------------------
python-jose[cryptography]==3.3.0    # JWT tokens
passlib[bcrypt]==1.7.4              # Password hashing
bcrypt==4.0.1
passlib==1.7.4                      # Data encryption

# -----------------------------------------
# 🤖 AI & Machine Learning
# -----------------------------------------
openai==1.31.0                       # OpenAI API
langchain==0.1.15                    # LangChain framework
langchain-openai==0.0.5             # OpenAI integration
scikit-learn==1.3.2                 # ML algorithms
pandas==2.1.4                       # Data analysis
numpy==1.26.2                       # Numerical computing

# -----------------------------------------
# 📄 File Processing
# -----------------------------------------
PyPDF2==3.0.1                       # PDF reading
reportlab==4.0.7                    # PDF generation
python-docx==1.1.0                  # Word document handling
Pillow==10.1.0                      # Image processing

# -----------------------------------------
# 🔧 Utilities
# -----------------------------------------
pydantic==2.5.0                     # Data validation
pydantic-settings==2.1.0            # Settings management
python-dotenv==1.0.0                # Environment variables
email-validator==2.1.0              # Email validation
aiofiles==23.2.1                    # Async file operations

# -----------------------------------------
# 📊 Data & Date Handling
# -----------------------------------------
python-dateutil==2.8.2              # Date utilities
pytz==2023.3                        # Timezone support

# -----------------------------------------
# 🧪 Testing (Optional)
# -----------------------------------------
pytest==7.4.3                       # Testing framework
pytest-asyncio==0.21.1              # Async testing
httpx==0.25.2                       # Async HTTP client
